dashboard/requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
sqlalchemy==2.0.36
mysql-connector-python==9.1.0
pandas==2.2.3
python-dotenv==1.0.1
